Ajout de l'état de partie "Non commencée"

+++ include/Case.h
@@ -24,7 +24,7 @@ namespace JeuHive {
 		const vector<const Piece*>& getPieces() const { return pieces; }
 
 		bool estVide() const { return pieces.empty(); }	// très peu pertinent car une case vide devrait être supprimée
-		int getNombrePieces() const { return pieces.size(); }
+		int getNombrePieces() const { return static_cast<int>(pieces.size()); }
 		Case(const Coordonnee& coo) : coo(coo) {}
 
 		string getString(int taille_str=8) const;
